Fixing up discrepancies
Here there have been some minor parameter differences that have led to very different results.

(1) Changing the deltaV to 0.65 makes the system tend toward Chaos, 0.7 seems to be too high for this to happen!

(2) Changing the initial conditions to be somewhere closer to what has been done previously -- this was needed

(3) Another suprising change was changing the Calcuim gain from 1.1 to 1.0, this is quite a sneaky difference. In all the papers that refernce the BTF model they all say they use the parameters from the paper, gCa=1.1 is in the table of the paper. However, gCa=1 is used in all versions of the code I have recieved!

(4) Changed the ode solver to ode23, this is for consistency, and also because 0de45 uses more points in the runge-kutta scheme, ode23 is a lower order runge-kutta scheme and having smaller number of points to interpolate the solution is making a difference.

(5) A major bug was cleaned up in this version -- i was updating the neural function incorrectly oops! This was the last little bit (and was not really the cause of the major discrepances as this was tested with the GUI on its own).